requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==35.0.0
Django==3.2.9
Flask==2.0.2
greenlet==1.1.2
h11==0.12.0
html5lib==1.1
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.2
lxml==4.6.4
MarkupSafe==2.0.1
numpy==1.21.4
outcome==1.1.0
pandas==1.3.4
prompt-toolkit==3.0.23
pycparser==2.21
pyee==8.2.2
pyOpenSSL==21.0.0
pyppeteer==0.2.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-http-client==3.3.4
pytz==2021.3
requests==2.26.0
selenium==4.0.0
sendgrid==6.9.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.27
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.5.2
pymysql
webencodings==0.5.1
websockets==9.1
Werkzeug==2.0.2
wsproto==1.0.0
zipp==3.6.0
